+++ src/LightParticle.java
@@ -70,6 +70,6 @@ public class LightParticle extends Particle {
     @Override
     public void draw(Graphics g) {
         // Отрисовка изображения текстуры частицы на графическом контексте
-        g.drawImage(texture, (int)x, (int)y, size, size, null);
+        g.drawImage(texture, x, y, size, size, null);
     }
 }
